article count, project count and research areas part is added

+++ research_portal/src/main/java/com/uni/research_portal/model/Project.java
@@ -9,7 +9,7 @@ public class Project {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "project_id")
-    private Long projectId;
+    private int projectId;
 
     @Column(nullable = false)
     private String projectName;
